Create a menu-burger.js component and import the const openBurgerMenu in the application.js file
