+++ app/assets/stylesheets/application.scss
@@ -11,16 +11,3 @@
 // Responsive design
 @import "responsive/tablet";
 @import "responsive/mobile";
-
-// Shared style
-body {
-  font-family: $sans-serif;
-  font-weight: $main-weight;
-  font-size: $main-size;
-}
-
-// Reset style
-* {
-  margin: 0px;
-  padding: 0px;
-}
