+++ app/javascript/components/navbar.js
@@ -1,0 +1,14 @@
+// const initUpdateNavbarOnScroll = () => {
+//   const navbar = document.querySelector('.navbar-lewagon');
+//   if (navbar) {
+//     window.addEventListener('scroll', () => {
+//       if (window.scrollY >= window.innerHeight) {
+//         navbar.classList.add('navbar-lewagon-white');
+//       } else {
+//         navbar.classList.remove('navbar-lewagon-white');
+//       }
+//     });
+//   }
+// }
+
+// export { initUpdateNavbarOnScroll };
